17.12. lbc playTheGame function added
