+++ HanngManOrWomanGame/index.html
@@ -42,20 +42,7 @@
       <button id="button">Z</button>
     </div>
 
-    <div id="word">
-      <!-- <div id="letterDiv">
-        <p id="letter">P</p>
-      </div>
-      <div id="letterDiv">
-        <p id="letter">R</p>
-      </div>
-      <div id="letterDiv">
-        <p id="letter">A</p>
-      </div>
-      <div id="letterDiv">
-        <p id="letter">A</p>
-      </div> -->
-    </div>
+    <div id="word"></div>
 
     <script type="module" src="index.js"></script>
   </body>
